Moving non-constants into utils. Renaming
