+++ sdk/monitor/azure-monitor-opentelemetry/azure/monitor/opentelemetry/_constants.py
@@ -4,14 +4,6 @@
 # license information.
 # --------------------------------------------------------------------------
 
-import logging
-import platform
-from os import environ
-from pathlib import Path
-
-from azure.monitor.opentelemetry.exporter._connection_string_parser import (  # pylint: disable=import-error,no-name-in-module
-    ConnectionStringParser,
-)
 from azure.monitor.opentelemetry.exporter._constants import (  # pylint: disable=import-error,no-name-in-module
     _AZURE_MONITOR_DISTRO_VERSION_ARG,
 )
@@ -35,58 +27,10 @@ SPAN_PROCESSORS_ARG = "span_processors"
 
 _LOG_PATH_LINUX = "/var/log/applicationinsights"
 _LOG_PATH_WINDOWS = "\\LogFiles\\ApplicationInsights"
-#TODO: move to utils. Then replace with exporter
-_IS_ON_APP_SERVICE = "WEBSITE_SITE_NAME" in environ
-#TODO: move to diagnostics
-# TODO: Add environment variable to enabled diagnostics off of App Service
-_IS_DIAGNOSTICS_ENABLED = _IS_ON_APP_SERVICE
-_CUSTOMER_IKEY_ENV_VAR = None
 _PREVIEW_ENTRY_POINT_WARNING = "Autoinstrumentation for the Azure Monitor OpenTelemetry Distro is in preview."
-logger = logging.getLogger(__name__)
 
 
-#TODO: move to utils or diagnostic logs
-# pylint: disable=global-statement
-def _get_customer_ikey_from_env_var():
-    global _CUSTOMER_IKEY_ENV_VAR
-    if not _CUSTOMER_IKEY_ENV_VAR:
-        _CUSTOMER_IKEY_ENV_VAR = "unknown"
-        try:
-            _CUSTOMER_IKEY_ENV_VAR = (
-                ConnectionStringParser().instrumentation_key
-            )
-        except ValueError as e:
-            logger.error("Failed to parse Instrumentation Key: %s", e)
-    return _CUSTOMER_IKEY_ENV_VAR
-
-
-#TODO: move to utils or diagnostic logs
-def _get_log_path(status_log_path=False):
-    system = platform.system()
-    if system == "Linux":
-        return _LOG_PATH_LINUX
-    if system == "Windows":
-        log_path = str(Path.home()) + _LOG_PATH_WINDOWS
-        if status_log_path:
-            return log_path + "\\status"
-        return log_path
-    return None
-
-
-#TODO: move to utils
-def _env_var_or_default(var_name, default_val=""):
-    try:
-        return environ[var_name]
-    except KeyError:
-        return default_val
-
-
-#TODO: move to utils
-_EXTENSION_VERSION = _env_var_or_default(
-    "ApplicationInsightsAgent_EXTENSION_VERSION", "disabled"
-)
-
-# Instrumentations
+# --------------------Instrumentations------------------------------
 
 # Opt-out
 _AZURE_SDK_INSTRUMENTATION_NAME = "azure_sdk"
