--- sdk/monitor/azure-monitor-opentelemetry/azure/monitor/opentelemetry/_util/__init__.py
@@ -1,5 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License in the project root for
-# license information.
-# --------------------------------------------------------------------------
