+++ sdk/monitor/azure-monitor-opentelemetry/tests/test_configure.py
